feat: 관리자만 강의 crud 가 가능하도록 수정
- 강의 등록, 조회는 관리자만 가능
- 강의 수정, 삭제는 MANAGER 만 가능

+++ src/main/java/com/example/lecturerestapi/service/LectureFacade.java
@@ -1,5 +1,7 @@
 package com.example.lecturerestapi.service;
 
+import com.example.lecturerestapi.entity.Admin;
+import com.example.lecturerestapi.entity.AdminRole;
 import com.example.lecturerestapi.entity.Category;
 import com.example.lecturerestapi.dto.LectureRequest;
 import com.example.lecturerestapi.dto.LectureResponse;
@@ -15,24 +17,39 @@ public class LectureFacade {
     private final LectureService lectureService;
     private final TeacherService teacherService;
 
-    public LectureResponse save(LectureRequest request) {
-        Teacher teacher = teacherService.findTeacherById(request.getTeacherId());
-        return lectureService.save(request, teacher);
+    public LectureResponse save(LectureRequest request, Admin admin) {
+        if (admin.getRole() == AdminRole.MANAGER || admin.getRole() == AdminRole.STAFF) {
+            Teacher teacher = teacherService.findTeacherById(request.getTeacherId());
+            return lectureService.save(request, teacher);
+        }
+        throw new IllegalArgumentException("관리자만 강사 등록이 가능합니다.");
     }
 
-    public LectureResponse update(Long lectureId, LectureRequest request) {
-        return lectureService.update(lectureId, request);
+    public LectureResponse update(Long lectureId, LectureRequest request, Admin admin) {
+        if (admin.getRole() == AdminRole.MANAGER) {
+            return lectureService.update(lectureId, request);
+        }
+        throw new IllegalArgumentException("MANAGER 만 강사 수정이 가능합니다.");
     }
 
-    public LectureResponse findById(Long lectureId) {
-        return lectureService.findById(lectureId);
+    public LectureResponse findById(Long lectureId, Admin admin) {
+        if (admin.getRole() == AdminRole.MANAGER || admin.getRole() == AdminRole.STAFF) {
+            return lectureService.findById(lectureId);
+        }
+        throw new IllegalArgumentException("관리자만 강사 조회가 가능합니다.");
     }
 
-    public List<LectureResponse> findByCategory(Category category) {
-        return lectureService.findByCategory(category);
+    public List<LectureResponse> findByCategory(Category category, Admin admin) {
+        if (admin.getRole() == AdminRole.MANAGER || admin.getRole() == AdminRole.STAFF) {
+            return lectureService.findByCategory(category);
+        }
+        throw new IllegalArgumentException("관리자만 강사 조회가 가능합니다.");
     }
 
-    public Long delete(Long lectureId) {
-        return lectureService.delete(lectureId);
+    public Long delete(Long lectureId, Admin admin) {
+        if (admin.getRole() == AdminRole.MANAGER) {
+            return lectureService.delete(lectureId);
+        }
+        throw new IllegalArgumentException("MANAGER 만 강사 삭제가 가능합니다.");
     }
 }
